+++ packages/target-tauri/src-tauri/src/settings.rs
@@ -33,6 +33,11 @@ pub(crate) const SYNC_ALL_ACCOUNTS_DEFAULT: bool = true;
 pub(crate) const THEME: &str = "activeTheme";
 pub(crate) const THEME_DEFAULT: &str = "system";
 
+pub(crate) const AUTOSTART_KEY: &str = "autostart";
+// IDEA: maybe we need to have more advanced logic for the default,
+// if we have other builds like portable builds for example
+pub(crate) const AUTOSTART_DEFAULT: bool = true;
+
 // runtime calls this when desktop settings change
 #[tauri::command]
 pub async fn change_desktop_settings_apply_side_effects(
@@ -51,6 +56,7 @@ pub async fn change_desktop_settings_apply_side_effects(
         // update "mute notification" menu item with new state
         NOTIFICATIONS => app.state::<TrayManager>().update_menu(&app).await,
         THEME => update_theme_in_other_windows(&app).context("update theme in other windows"),
+        AUTOSTART_KEY => apply_autostart(&app),
         _ => Ok(()),
     }
     .map_err(|err| format!("{err:#}"))
@@ -64,6 +70,7 @@ pub(crate) async fn load_and_apply_desktop_settings_on_startup(
     app.state::<TrayManager>()
         .apply_wanted_active_state(app)
         .await?;
+    apply_autostart(&app)?;
 
     Ok(())
 }
@@ -173,4 +180,25 @@ impl<R: tauri::Runtime> StoreExtBoolExt for tauri_plugin_store::Store<R> {
             .and_then(|v| v.as_bool())
             .unwrap_or(default_value)
     }
+}
+
+#[cfg(not(desktop))]
+pub(crate) fn apply_autostart(app: &AppHandle) -> anyhow::Result<()> {
+    Ok(())
+}
+
+#[cfg(desktop)]
+pub(crate) fn apply_autostart(app: &AppHandle) -> anyhow::Result<()> {
+    use tauri_plugin_autostart::ManagerExt;
+    let store = app.store(CONFIG_FILE)?;
+    let enabled = get_setting_bool_or(store.get(AUTOSTART_KEY), AUTOSTART_DEFAULT);
+
+    let autostart_manager = app.autolaunch();
+
+    if enabled {
+        autostart_manager.enable()?;
+    } else {
+        autostart_manager.disable()?;
+    }
+    Ok(())
 }